Add recovery script
This was used to recover data from a backup I had made using backme.
Unfortunately, I had some trouble (hence the commented-out lines all
over), and I have decided to forego this backup scheme. This script
will not work as-is, but should work fine for small backups with a
few changes. I had to do some bizzare things to recover my data and
am putting this here for historical purposes in case I need it
again.